Simulate event bubbling on detached event targets
When we used jQuery to trigger events, events always bubbled regardless
of the attachment status of the target element. This commit preserves
that behavior with native DOM events.

I am a bit worried about doing the document.contains check on every
keystroke. We may want a flag that only enables this feature in specs.