Add 'abort!' command which terminates search for a matching binding
atom/vim-mode#348 reveals a hole in our current binding system.

@jamesgecko wants to bind a certain command both to .workspace and also
to `.editor:not(.insert-mode)`. This would allow the same binding to
function in editors that aren't in insert mode and also when no editors
are open at all. Unfortunately, it doesn't work because in insert mode,
the event bubbles through the editor without matching but always matches
the `.workspace` selector.

This commit adds the best feature I can think of to address this
situation. Now the user can just add the binding to `.workspace`, then
exclude insert mode by adding the same binding to `.editor.insert-mode`
with an `abort!` directive.