Always reload keymaps if CSON parsing of file succeeded
If CSON parsing succeeded, but the file was empty or contained only
comments -- the return value of readKeymap will be null, so the check
for whether to reload or not will fail. This tweaks the check to not
reload keymaps only if the result was undefined, which happens only
if CSON parsing failed.