Don’t shadow ASCII characters with alt-bindings
If alt- or shift-alt- produces an ASCII character on the current layout,
interpret the keystroke as that character. If ctrl or cmd are also 
pressed, don’t use the alternative character for the binding.

For example, on a Swiss layout:
alt-g is interpreted as @
ctrl-alt-g is still interpreted as ctrl-alt-g