Use KeyboardEvent.key property as much as possible
The only time that we can't use KeyboardEvent.key is on
macOS when the option key is held down. In that case,
the .key property will contain an alternative character.
In order to match alt bindings, we need to know the character
that the key would have typed if option were not held down.
We determine this character using the keyboard-layout module.

Signed-off-by: Nathan Sobo <nathan@github.com>