Always favor AltGraph variants on Windows
Unlike macOS, Windows doesn't always have an AltGraph variant for every key in every layout. So when one does exist, users really expect for it to work. When a conflicting ctrl-alt- binding exists, it won't be possible to use it and the user will need to create a new binding.