*Always* use U.S. layout chars when modifiers are used on non-Latin KBs
Previously, we only translated non-Latin characters to their U.S.
equivalent, but on some keyboards such as Russian and Hebrew, latin
symbols are in different locations and users expect key bindings to use
the U.S. location for these symbols.

This commit checks entire *layouts* for being non-Latin by looking for
non-Latin characters in the first 4 keys of the home row rather than
testing characters on an individual basis. If the keyboard layout is
detected to be non-Latin, we fall back to the U.S. layout for bindings
across the board.