Don’t resolve dead keys on Windows
This avoids a bug in Windows where dead key state is blown away if we
attempt to get the character corresponding to a dead key. We now skip
dead keys in getCurrentkeymap on Windows to avoid this issue, making it
impossible to resolve dead keys to actual characters for the purposes
of key bindings.