Restore ability to bind dead keys on macOS
A previous fix for IME input bails out of keystroke handling if the
legacy keyCode property is 229. Unfortunately, this is also the value
of keyCode for keystrokes like alt-u, which don’t pop up the IME menu.
These keystrokes should be handled normally. It seems like the best way
to distinguish these keys from the IME case is that their key value is
‘Dead’.